Remove github-pages gem version lock
For some reason, the github-pages was locked on version 203 in the
Gemfile. I don't know why this is, but it seems that it was locked on
that version by the boilerplate.

I updated the gem because a major security vulnerability was generated
for the kramdown dependency. But by unlocking the gem, we should get a
lot of other cool updates! Awesome!